refactor: moved pixelRatio stuff
